Meanwhile I had further thoughts about the implementation.  I realised one of the 'computeP's could be delayed (for b' as it can be done while computing b'').
This should improve inlining.

In fact I thought r'' may not need to be computed until the final
iteration, but such a change may mean delaying a computation which
requires access to all the r' from every iteration - probably not a
good idea.  However it seems counter intuitive to me that use of omega
on red cells has no influence on black cell calculations.  It just
affects the very last computation of red cells.

I will investigate just using omega for black cells and see if it works as well.

Here is the version with b' delayed.